+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [Component height can be resized][14849]
 - [Add right click menu for multiple components][14640].
 - [Execution can be scheduled for the specific version tag][14883]
+- [Add component spacing options][14888]
 
 [14590]: https://github.com/enso-org/enso/pull/14590
 [14678]: https://github.com/enso-org/enso/pull/14678
@@ -26,6 +27,7 @@
 [14849]: https://github.com/enso-org/enso/pull/14849
 [14640]: https://github.com/enso-org/enso/pull/14640
 [14883]: https://github.com/enso-org/enso/pull/14883
+[14888]: https://github.com/enso-org/enso/pull/14888
 
 #### Enso Standard Library
 
@@ -44,6 +46,8 @@
 - [New `Profile` API for timing code execution.][14827]
 - [Expanded S3 API with versions and signed uri][14831]
 - [Support for reading JSON data from database connections.][14872]
+- [Add Redshift bulk loading][14860]
+- [JSON functions on Table and Column. XML and Hashset fixes.][14884]
 
 [14522]: https://github.com/enso-org/enso/pull/14522
 [14476]: https://github.com/enso-org/enso/pull/14476
@@ -60,6 +64,8 @@
 [14827]: https://github.com/enso-org/enso/pull/14827
 [14831]: https://github.com/enso-org/enso/pull/14831
 [14872]: https://github.com/enso-org/enso/pull/14872
+[14860]: https://github.com/enso-org/enso/pull/14860
+[14884]: https://github.com/enso-org/enso/pull/14884
 
 #### Enso Language & Runtime
 
@@ -69,6 +75,7 @@
 - [Flush system caches via `Runtime.gc`][14557]
 - [Removing `catch_primitive` from the API][14676]
 - [`polyglot java import` loads classes from HotSpot JVM][14798]
+- [Require opt-in to prefer local libraries][14885]
 
 [14480]: https://github.com/enso-org/enso/pull/14480
 [14490]: https://github.com/enso-org/enso/pull/14490
@@ -76,6 +83,7 @@
 [14557]: https://github.com/enso-org/enso/pull/14557
 [14676]: https://github.com/enso-org/enso/pull/14676
 [14798]: https://github.com/enso-org/enso/pull/14798
+[14885]: https://github.com/enso-org/enso/pull/14885
 
 # Enso 2025.3